47.1.Setting the UUID on POST
The UUID is now the identifier inside of our User resource:

 306 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @ApiProperty(identifier=true)
     */
    private $uuid;
}
Awesome! But it still works exactly like the old ID. What I mean is, only the server can set the UUID. If we tried to send UUID as a JSON field when creating a user, it would be ignored.
How can I be so sure? Well, look at the User class: $uuid is not settable anywhere. It's not an argument to the constructor and there's no setUuid() method
Time to change that!

Setting the UUID in a Test
Let's describe the behavior we want in a test. In UserResourceTest, go up to the top and copy testCreateUser(). Paste that down here and call it testCreateUserWithUuid()
The key change we want to make is this: in the JSON, we're going to pass a uuid field. For the value, go up and say $uuid = Uuid - the one from Ramsey - ::uuid4(). Then below, send that as the uuid
I technically could call ->toString()... but since the Uuid object has an __toString() method, we don't need to. Assert that the response is a 201 and... then we can remove the part that fetches the User from the database. Because... we know that the @id should be /api/users/ and then that . I'll also remove the login part, only because we have that in the other test
So this is the plan: we send the uuid and it uses that uuid. Copy the name of this method and let's make sure it fails:

  symfony php bin/phpunit --filter=testCreateUserWithUuid
It does. It completely ignores the UUID that we send and generates its own.