47.2.Setting the UUID on POST
Making the uuid Field Settable
So how can we make the UUID field settable? Well, it's really no different than any other field: we need to put the property in the correct group and make sure it's settable either through the constructor or via a setter method.
Let's think: we only want this field to be settable on create: we don't want to allow anyone to modify it later. So we could add a setUuid() method, but then we would need to be careful to configure and add the correct groups so that it can be set on create but not edit.
But... there's a simpler solution: avoid the setter and instead add  as an argument to the constructor! Then, by the rules of object-oriented coding, it will be settable on create but immutable after.
Let's do that: add a UuidInterface  argument and default it to null. Then $this->uuid = $uuid ?: Uuid::uuid4()
So if a $uuid argument is passed, we'll use that. If not, we generate a new one. Oh, and we also need to make sure the UUID is actually writeable in the API. Above the  property, add @Groups() with user:write
Ok, let's try the test again!

symfony php bin/phpunit --filter=testCreateUserWithUuid
This time... woh! It works. That's awesome. And the documentation for this instantly looks perfect. Refresh the API homepage, open up the POST operation for users, hit "Try it out" and... yep! It already shows a UUID example and it understands that it is available for us to set.

UUID String Transformed to an Object?
But wait a second. How did that work? Think about it, if you look at our test, we're sending a string in the JSON:

 110 lines  tests/Functional/UserResourceTest.php
class UserResourceTest extends CustomApiTestCase
{
    public function testCreateUserWithUuid()
    {
        $uuid = Uuid::uuid4();
        $client->request('POST', '/api/users', [
            'json' => [
                'uuid' => $uuid,
            ]
        ]);
    }
}
But ultimately, on our User object, the constructor argument accepts a UuidInterface object, not a string:

 307 lines  src/Entity/User.php
class User implements UserInterface
{
    public function __construct(UuidInterface $uuid = null)
    {
    }
}
How did that string become an object?

Remember: API platform - well really, Symfony's serializer - is really good at reading your types. It notices that the type for $uuid is UuidInterface and uses that to try to find a denormalizer that understands this type. And fortunately, API Platform comes with a denormalizer that works with ramsey UUID's out of the box. Yep, that denormalizer takes the string and turns it into a UUID object so that it can then be passed to the constructor.
So... yay UUIDs! But, before we finish, there is one tiny quirk with UUID's. Let's see what it is next and learn how to work around it.