48.1.UUID Quirk with "id" Name
There is one tiny little quirk with you UUID's. What is it? If you want to be able to send it in JSON, the field can't be called... id!

Naming the Identifier "id"
I know, that's sounds kinda strange, so let me show you. Find the uuid property and pretend that we want to call this id in the API:

 307 lines  src/Entity/User.php
class User implements UserInterface
{
    /**
     * @ORM\Column(type="uuid", unique=true)
     * @ApiProperty(identifier=true)
     * @Groups({"user:write"})
     */
    private \uuid;
}
Literally, instead of sending a field called uuid, I want to send one called id.
The easiest way to make that happen is to add @SerializedName("id")
Then over in the test, it's simple: change uuid to id
We've done this type of thing before. And until now, it's been working fine. If you're getting the feeling that it won't work this time... yea... you're right. Run that test again:

  symfony php bin/phpunit --filter=testCreateUserWithUuid
It fails! It says:

  Update is not allowed for this operation.

This... is sort of a bug in API Platform. Well... it's not that simple - it's related to the idea that the id field is, sort of special, in the jsonapi format.
Obviously, the easiest fix is to not call your field id if you want to allow it to be sent on create. But, if you really do want to call it id, then we have two work-arounds.