48.3.UUID Quirk with "id" Name
Removing the JSON format Entirely
If you're the only person using your API, then this is probably ok: you just need to know to include this specific header. But if third-parties will use your API, then it's kind of ugly to force them to pass this exact header or get a weird error.
Oh, and by the way, your API clients do need to include a Content-Type header when sending data, otherwise API Platform won't understand it or will think that you're sending form data instead of JSON. Basically, you'll get a 400 error. But the Content-Type header is usually sent automatically by most clients and set to application/json... because the client realizes you're sending JSON.
So... if you want to prevent your API users from needing to override that and send application/ld+json, there is another - kind of more extreme - option: disable the JSON format and always force JSON-LD.
Let's try this. Remove the header
And then go into config/packages/api_platform.yaml. Down here, comment-out the json format completely
This means that this is no longer a format that our API can read or output.
There's one other spot I need to change to avoid an error. Open CheeseListing. In a previous tutorial, we set specific formats for this resource to allow a csv format. Remove json from this list
Ok, let's see what happens! When we run the test...

  symfony php bin/phpunit --filter=testCreateUserWithUuid
It still fails. And the error is cool!
The content-type "application/json" is not supported
The test client - realizing that we're sending JSON data - sent this header automatically for us. But now that our API doesn't support JSON, this fails!