48.4.UUID Quirk with "id" Name
Back in api_platform.yaml, move that application/json line up under the jsonld mime type
This means that if the Content-Type header is application/json, use jsonld.

Try the test one more time.

  symfony php bin/phpunit --filter=testCreateUserWithUuid
And... got it! This solution is a bit odd... but if you're always using JSON-LD, it's one option.